move pageselector compontent to a component folder
