+++ client/src/components/PageSelector.tsx
